Use the new userboot 'getenv' callback to set a couple of environment variables in the guest.
The variables are: smbios.bios.vendor=BHYVE and boot_serial=1

The FreeBSD guest uses the "smbios.bios.vendor" environment variable to
detect whether or not it is running as a guest inside a hypervisor.

The "boot_serial=1" is temporary and will be dropped when bhyve can do VGA
emulation.

Obtained from:	NetApp


git-svn-id: svn://svn.freebsd.org/base/projects/bhyve@242676 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f